Refactor file structure for generator sdk
