+++ src/rust/rust-proto-new/src/graplinc/grapl/api/plugin_sdk/generators/v1beta1/server.rs
@@ -1,4 +1,7 @@
-use std::time::Duration;
+use std::{
+    marker::PhantomData,
+    time::Duration,
+};
 
 use futures::{
     channel::oneshot::{
@@ -86,6 +89,7 @@ where
     shutdown_rx: Receiver<()>,
     service_name: &'static str,
     identity: Identity,
+    f_: PhantomData<F>,
 }
 
 impl<T, H, F> GeneratorServer<T, H, F>
@@ -116,6 +120,7 @@ where
                 shutdown_rx,
                 service_name: GeneratorServiceProto::<GrpcApi<T>>::NAME,
                 identity,
+                f_: PhantomData,
             },
             shutdown_tx,
         )
