+++ src/rust/rust-proto-new/src/lib.rs
@@ -143,8 +143,6 @@ pub mod graplinc {
             pub mod plugin_sdk {
                 pub mod generators {
                     pub mod v1beta1;
-                    mod v1beta1_client;
-                    mod v1beta1_server;
                 }
             }
